move more wallet connect state
