+++ src/ConnectWalletButton.tsx
@@ -1,27 +1,20 @@
-import { useState } from 'react';
+import React from 'react';
 import styles from './ConnectWalletButton.module.css';
 import { formatWalletAddress } from './utils/helpers';
 import WalletModal from './WalletModal';
+import { useWalletState } from './stores/walletStore/useWalletState';
 
-interface ConnectWalletButtonProps {
-  walletAddress: string;
-  handleConnectClick: () => void;
-  disconnectWallet: () => void;
-  icon?: string;
-  walletName?: string;
-  showSwitchWallet: boolean;
-}
-
-const ConnectWalletButton = ({
-  walletAddress,
-  handleConnectClick,
-  disconnectWallet,
-  icon,
-  walletName,
-  showSwitchWallet,
-}: ConnectWalletButtonProps) => {
-  const [isWalletModalOpen, setIsWalletModalOpen] = useState(false);
-  const [isDropdownOpen, setIsDropdownOpen] = useState(false);
+const ConnectWalletButton = () => {
+  const {
+    walletAddress,
+    walletProviderInfo,
+    availableProviders,
+    refreshProviders,
+    disconnectWallet,
+  } = useWalletState();
+  const { icon, name: walletName } = walletProviderInfo || {};
+  const [isWalletModalOpen, setIsWalletModalOpen] = React.useState(false);
+  const [isDropdownOpen, setIsDropdownOpen] = React.useState(false);
 
   const displayedButtonText = walletAddress
     ? formatWalletAddress(walletAddress)
@@ -38,12 +31,13 @@ const ConnectWalletButton = ({
   };
 
   const onSwitchWalletClick = () => {
-    handleConnectClick();
+    setIsWalletModalOpen(true);
     setIsDropdownOpen(false);
   };
 
-  const onConnectClick = () => {
+  const onButtonClick = () => {
     if (!walletAddress) {
+      refreshProviders();
       setIsWalletModalOpen(true);
     }
   };
@@ -56,7 +50,7 @@ const ConnectWalletButton = ({
       onMouseEnter={() => walletAddress && setIsDropdownOpen(true)}
       onMouseLeave={() => setIsDropdownOpen(false)}
     >
-      <button onClick={onConnectClick} className={styles.walletButton}>
+      <button onClick={onButtonClick} className={styles.walletButton}>
         {showWalletIcon && (
           <img
             src={icon}
@@ -72,7 +66,7 @@ const ConnectWalletButton = ({
           <div className={styles.dropdownItem} onClick={copyAddressToClipboard}>
             Copy Address
           </div>
-          {showSwitchWallet && (
+          {availableProviders.length > 1 && (
             <div className={styles.dropdownItem} onClick={onSwitchWalletClick}>
               Switch Wallet
             </div>
