+++ src/stores/walletStore/useWalletState.ts
@@ -1,6 +1,10 @@
 import React from 'react';
 import { ChatMessage, WalletInfo, WalletProviderDetail } from '../../types';
-import { EIP6963ProviderDetail, WalletProvider, walletStore } from './walletStore';
+import {
+  EIP6963ProviderDetail,
+  WalletProvider,
+  walletStore,
+} from './walletStore';
 import {
   formatWalletBalancesForPrompt,
   getChainAccounts,
@@ -26,6 +30,52 @@ export const useWalletState = () => {
     pendingWalletMessage: null,
   });
 
+  // const addWalletSystemMessage = useCallback(
+  //   async (walletInfo?: WalletInfo) => {
+  //     try {
+  //       let messageContent = '';
+
+  //       //  If no wallet info was provided, then we have been disconnected,
+  //       //  Reset the ref
+  //       if (!walletInfo) {
+  //         messageContent =
+  //           'Wallet has been disconnected. All previous wallet information is no longer valid.';
+  //         walletUpdateRef.current.previousAddress = '';
+  //         walletUpdateRef.current.previousChainId = '';
+  //       } else if (
+  //         walletInfo.address === walletUpdateRef.current.previousAddress &&
+  //         walletInfo.chainId === walletUpdateRef.current.previousChainId
+  //       ) {
+  //         return;
+  //       } else {
+  //         messageContent = `Wallet account changed. New address: ${walletInfo.address} on chain ID: ${walletInfo.chainId}.
+  //       Wallet type: ${walletInfo.walletType}.
+  //       Keep previous wallet information in context, but recognize that it is not current. ${walletInfo.balancesPrompt}`;
+
+  //         walletUpdateRef.current.previousAddress = walletInfo.address;
+  //         walletUpdateRef.current.previousChainId = walletInfo.chainId;
+  //       }
+
+  //       const walletMessage: ChatMessage = {
+  //         role: 'system',
+  //         content: messageContent,
+  //       };
+
+  //       //  If the conversation has already started, add the wallet message
+  //       if (activeChat.messageHistoryStore.getSnapshot().length > 0) {
+  //         activeChat.messageHistoryStore.addMessage(walletMessage);
+  //       } else {
+  //         //  Otherwise, store it in the ref to be added during handleChatCompletion
+  //         //  after the initial system prompt (so as not to override it)
+  //         walletUpdateRef.current.pendingWalletMessage = walletMessage;
+  //       }
+  //     } catch (error) {
+  //       console.error('Failed to add wallet system message:', error);
+  //     }
+  //   },
+  //   [activeChat],)
+  // );
+
   const disconnectWallet = React.useCallback(async () => {
     walletUpdateRef.current.isProcessing = true;
 
@@ -37,10 +87,20 @@ export const useWalletState = () => {
   }, []);
 
   // Wallet providers //////////////////////////////
-
   const [availableProviders, setAvailableProviders] = React.useState<
     WalletProviderDetail[]
-  >([]);
+  >(walletStore.getProviders());
+
+  const walletProviderInfo = React.useMemo(() => {
+    return walletConnection.isWalletConnected && walletConnection.rdns
+      ? availableProviders.find((p) => p.info.rdns === walletConnection.rdns)
+          ?.info
+      : undefined;
+  }, [
+    availableProviders,
+    walletConnection.isWalletConnected,
+    walletConnection.rdns,
+  ]);
 
   const refreshProviders = React.useCallback(() => {
     setAvailableProviders(walletStore.getProviders());
@@ -112,11 +172,66 @@ export const useWalletState = () => {
     [connectEIP6963Provider],
   );
 
+  //  Subscribe to wallet connection changes and update system prompt accordingly
+  React.useEffect(() => {
+    //  Only proceed if we're not already processing an update
+    if (walletUpdateRef.current.isProcessing) {
+      return;
+    }
+
+    //  Set up subscription for wallet changes
+    const unsubscribe = walletStore.subscribe(async () => {
+      const currentState = walletStore.getSnapshot();
+
+      //  Only process if we're not already handling a wallet update
+      if (walletUpdateRef.current.isProcessing) {
+        return;
+      }
+
+      try {
+        if (
+          walletUpdateRef.current.previousAddress ||
+          walletUpdateRef.current.previousChainId
+        ) {
+          if (currentState.isWalletConnected && currentState.provider) {
+            const addressChanged =
+              currentState.walletAddress !==
+              walletUpdateRef.current.previousAddress;
+            const chainChanged =
+              currentState.walletChainId !==
+              walletUpdateRef.current.previousChainId;
+
+            if (addressChanged || chainChanged) {
+              walletUpdateRef.current.isProcessing = true;
+              const walletInfo = await getCurrentWalletInfoForPrompt();
+              // await addWalletSystemMessage(walletInfo);
+            }
+          }
+          // Wallet was disconnected from outside our UI and disconnectWallet()
+          // wasn't called (which would have set isProcessing=true)
+          else if (
+            !currentState.isWalletConnected &&
+            walletUpdateRef.current.previousAddress.length > 0
+          ) {
+            walletUpdateRef.current.isProcessing = true;
+            // await addWalletSystemMessage();
+          }
+        }
+      } finally {
+        walletUpdateRef.current.isProcessing = false;
+      }
+    });
+
+    //  Clean up subscription
+    return () => unsubscribe();
+  }, [getCurrentWalletInfoForPrompt]);
+
   return {
     disconnectWallet,
     walletAddress: walletConnection.walletAddress,
     availableProviders,
     refreshProviders,
     handleProviderSelect,
+    walletProviderInfo,
   };
 };
