+++ src/useChat.tsx
@@ -1,4 +1,4 @@
-import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import { useCallback, useEffect, useState } from 'react';
 import { MessageHistoryStore } from './stores/messageHistoryStore';
 import {
   ActiveChat,
@@ -6,8 +6,6 @@ import {
   ConversationHistories,
   ConversationHistory,
   SearchableChatHistories,
-  WalletInfo,
-  WalletProviderDetail,
 } from './types';
 import {
   deleteConversation,
@@ -25,17 +23,9 @@ import { doChat, generateConversationTitle } from './api/chat';
 import { initializeToolCallRegistry } from './toolcalls/chain';
 import { ToolCallStreamStore } from './stores/toolCallStreamStore';
 import { useExecuteToolCall } from './useExecuteToolCall';
-import {
-  WalletProvider,
-  EIP6963ProviderDetail,
-  walletStore,
-} from './stores/walletStore';
+import { walletStore } from './stores/walletStore';
 import { defaultSystemPrompt } from './toolcalls/chain/prompts';
-import { useWalletStore } from './stores/walletStore/useWalletStore';
-import {
-  formatWalletBalancesForPrompt,
-  getChainAccounts,
-} from './utils/wallet';
+
 import { ModelId } from './types/index.ts';
 
 export const USE_LITELLM_TOKEN =
@@ -55,13 +45,6 @@ export function getToken() {
 }
 
 const activeChats: Record<string, ActiveChat> = {};
-
-interface WalletUpdateRef {
-  isProcessing: boolean;
-  previousAddress: string;
-  previousChainId: string;
-  pendingWalletMessage: ChatMessage | null;
-}
 
 export const useChat = (initValues?: ChatMessage[], initModel?: string) => {
   const [client] = useState(() => {
@@ -94,199 +77,6 @@ export const useChat = (initValues?: ChatMessage[], initModel?: string) => {
 
   const [toolCallRegistry] = useState(() => initializeToolCallRegistry());
 
-  const walletConnection = useWalletStore(walletStore);
-  const walletAddress = walletConnection.walletAddress;
-
-  const [availableProviders, setAvailableProviders] = useState<
-    WalletProviderDetail[]
-  >([]);
-
-  const [isWalletModalOpen, setIsWalletModalOpen] = useState(false);
-  const [isWalletConnecting, setIsWalletConnecting] = useState(false);
-
-  //  Ref to track async wallet switching and store pending wallet message
-  const walletUpdateRef = useRef<WalletUpdateRef>({
-    isProcessing: false,
-    previousAddress: '',
-    previousChainId: '',
-    pendingWalletMessage: null,
-  });
-
-  const getCurrentWalletInfoForPrompt = useCallback(async () => {
-    const walletConnection = walletStore.getSnapshot();
-
-    const walletInfo: WalletInfo = {
-      isConnected: walletConnection.isWalletConnected,
-      address: walletConnection.walletAddress,
-      chainId: walletConnection.walletChainId,
-      balancesPrompt: '',
-      walletType: walletConnection.walletType,
-    };
-
-    if (walletInfo.isConnected && walletConnection.provider) {
-      const balances = await getChainAccounts(walletConnection.provider);
-      walletInfo.balancesPrompt = formatWalletBalancesForPrompt(
-        balances,
-        walletConnection.walletChainId,
-      );
-    }
-
-    return walletInfo;
-  }, []);
-
-  const addWalletSystemMessage = useCallback(
-    async (walletInfo?: WalletInfo) => {
-      try {
-        let messageContent = '';
-
-        //  If no wallet info was provided, then we have been disconnected,
-        //  Reset the ref
-        if (!walletInfo) {
-          messageContent =
-            'Wallet has been disconnected. All previous wallet information is no longer valid.';
-          walletUpdateRef.current.previousAddress = '';
-          walletUpdateRef.current.previousChainId = '';
-        } else if (
-          walletInfo.address === walletUpdateRef.current.previousAddress &&
-          walletInfo.chainId === walletUpdateRef.current.previousChainId
-        ) {
-          return;
-        } else {
-          messageContent = `Wallet account changed. New address: ${walletInfo.address} on chain ID: ${walletInfo.chainId}.
-        Wallet type: ${walletInfo.walletType}.
-        Keep previous wallet information in context, but recognize that it is not current. ${walletInfo.balancesPrompt}`;
-
-          walletUpdateRef.current.previousAddress = walletInfo.address;
-          walletUpdateRef.current.previousChainId = walletInfo.chainId;
-        }
-
-        const walletMessage: ChatMessage = {
-          role: 'system',
-          content: messageContent,
-        };
-
-        //  If the conversation has already started, add the wallet message
-        if (activeChat.messageHistoryStore.getSnapshot().length > 0) {
-          activeChat.messageHistoryStore.addMessage(walletMessage);
-        } else {
-          //  Otherwise, store it in the ref to be added during handleChatCompletion
-          //  after the initial system prompt (so as not to override it)
-          walletUpdateRef.current.pendingWalletMessage = walletMessage;
-        }
-      } catch (error) {
-        console.error('Failed to add wallet system message:', error);
-      }
-    },
-    [activeChat],
-  );
-
-  const connectEIP6963Provider = useCallback(
-    async (rdns: string, chainId?: string) => {
-      if (walletUpdateRef.current.isProcessing) return;
-      walletUpdateRef.current.isProcessing = true;
-
-      try {
-        await walletStore.connectWallet({
-          chainId,
-          walletProvider: WalletProvider.EIP6963,
-          rdns,
-        });
-
-        const walletInfo = await getCurrentWalletInfoForPrompt();
-
-        //  Add a message if connection was successful
-        if (walletInfo.isConnected && walletInfo.address) {
-          await addWalletSystemMessage(walletInfo);
-        }
-      } catch (error) {
-        console.error('Failed to connect wallet:', error);
-        throw error;
-      } finally {
-        walletUpdateRef.current.isProcessing = false;
-      }
-    },
-    [getCurrentWalletInfoForPrompt, addWalletSystemMessage],
-  );
-
-  //  Subscribe to wallet connection changes and update system prompt accordingly
-  useEffect(() => {
-    //  Only proceed if we're not already processing an update
-    if (walletUpdateRef.current.isProcessing) {
-      return;
-    }
-
-    //  Set up subscription for wallet changes
-    const unsubscribe = walletStore.subscribe(async () => {
-      const currentState = walletStore.getSnapshot();
-
-      //  Only process if we're not already handling a wallet update
-      if (walletUpdateRef.current.isProcessing) {
-        return;
-      }
-
-      try {
-        if (
-          walletUpdateRef.current.previousAddress ||
-          walletUpdateRef.current.previousChainId
-        ) {
-          if (currentState.isWalletConnected && currentState.provider) {
-            const addressChanged =
-              currentState.walletAddress !==
-              walletUpdateRef.current.previousAddress;
-            const chainChanged =
-              currentState.walletChainId !==
-              walletUpdateRef.current.previousChainId;
-
-            if (addressChanged || chainChanged) {
-              walletUpdateRef.current.isProcessing = true;
-              const walletInfo = await getCurrentWalletInfoForPrompt();
-              await addWalletSystemMessage(walletInfo);
-            }
-          }
-          // Wallet was disconnected from outside our UI and disconnectWallet()
-          // wasn't called (which would have set isProcessing=true)
-          else if (
-            !currentState.isWalletConnected &&
-            walletUpdateRef.current.previousAddress.length > 0
-          ) {
-            walletUpdateRef.current.isProcessing = true;
-            await addWalletSystemMessage();
-          }
-        }
-      } finally {
-        walletUpdateRef.current.isProcessing = false;
-      }
-    });
-
-    //  Clean up subscription
-    return () => unsubscribe();
-  }, [getCurrentWalletInfoForPrompt, addWalletSystemMessage]);
-
-  const refreshProviders = useCallback(() => {
-    setAvailableProviders(walletStore.getProviders());
-  }, []);
-
-  const openWalletConnectModal = useCallback(() => {
-    refreshProviders();
-    setIsWalletModalOpen(true);
-  }, [refreshProviders, setIsWalletModalOpen]);
-
-  const handleProviderSelect = useCallback(
-    async (provider: EIP6963ProviderDetail) => {
-      try {
-        await connectEIP6963Provider(
-          provider.info.rdns,
-          `0x${Number(2222).toString(16)}`,
-        );
-      } catch (err) {
-        console.error(
-          `Failed to connect to ${provider.info.name}: ${(err as Error).message}`,
-        );
-      }
-    },
-    [connectEIP6963Provider],
-  );
-
   const setIsOperationValidated = useCallback(
     (isOperationValidated: boolean) => {
       setActiveChat((prev) => {
@@ -297,29 +87,15 @@ export const useChat = (initValues?: ChatMessage[], initModel?: string) => {
     [],
   );
 
-  const { executeOperation, handleModalClose } = useExecuteToolCall(
+  const { executeOperation } = useExecuteToolCall(
     toolCallRegistry,
     walletStore,
     activeChat.isOperationValidated,
     setIsOperationValidated,
-    openWalletConnectModal,
-    isWalletConnecting,
-    setIsWalletConnecting,
-  );
-
-  const closeWalletConnectModal = useCallback(() => {
-    setIsWalletModalOpen(false);
-    if (handleModalClose) {
-      handleModalClose();
-    }
-  }, [handleModalClose]);
-
-  const onProviderSelect = useCallback(
-    async (provider: WalletProviderDetail) => {
-      await handleProviderSelect(provider);
-      closeWalletConnectModal();
-    },
-    [handleProviderSelect, closeWalletConnectModal],
+    // TODO: Add these back in when we have a way to open the wallet connect modal
+    // openWalletConnectModal,
+    // isWalletConnecting,
+    // setIsWalletConnecting,
   );
 
   const fetchConversations = useCallback(() => {
@@ -358,13 +134,14 @@ export const useChat = (initValues?: ChatMessage[], initModel?: string) => {
           content: defaultSystemPrompt,
         });
 
+        // TODO: handle wallet connect system messages
         // Add the pending wallet message if it exists
-        if (walletUpdateRef.current.pendingWalletMessage) {
-          newActiveChat.messageHistoryStore.addMessage(
-            walletUpdateRef.current.pendingWalletMessage,
-          );
-          walletUpdateRef.current.pendingWalletMessage = null;
-        }
+        // if (walletUpdateRef.current.pendingWalletMessage) {
+        //   newActiveChat.messageHistoryStore.addMessage(
+        //     walletUpdateRef.current.pendingWalletMessage,
+        //   );
+        //   walletUpdateRef.current.pendingWalletMessage = null;
+        // }
       }
 
       // Add new messages to history
@@ -462,32 +239,33 @@ export const useChat = (initValues?: ChatMessage[], initModel?: string) => {
 
     setActiveChat(newChat);
 
+    // TODO: handle wallet connect system messages
     // If wallet is connected, add the info as a separate system message
-    try {
-      walletUpdateRef.current.isProcessing = true;
+    // try {
+    //   walletUpdateRef.current.isProcessing = true;
 
-      const walletInfo = await getCurrentWalletInfoForPrompt();
+    //   const walletInfo = await getCurrentWalletInfoForPrompt();
 
-      if (walletInfo.isConnected && walletInfo.address) {
-        const walletInfoMessage: ChatMessage = {
-          role: 'system',
-          content: `Current wallet information: Address: ${walletInfo.address} on chain ID: ${walletInfo.chainId}.
-          Wallet type: ${walletInfo.walletType || 'Unknown'}.
-          ${walletInfo.balancesPrompt}`,
-        };
+    //   if (walletInfo.isConnected && walletInfo.address) {
+    //     const walletInfoMessage: ChatMessage = {
+    //       role: 'system',
+    //       content: `Current wallet information: Address: ${walletInfo.address} on chain ID: ${walletInfo.chainId}.
+    //       Wallet type: ${walletInfo.walletType || 'Unknown'}.
+    //       ${walletInfo.balancesPrompt}`,
+    //     };
 
-        newChat.messageHistoryStore.addMessage(walletInfoMessage);
+    //     newChat.messageHistoryStore.addMessage(walletInfoMessage);
 
-        //  Update ref with current wallet info
-        walletUpdateRef.current.previousAddress = walletInfo.address;
-        walletUpdateRef.current.previousChainId = walletInfo.chainId;
-      }
-    } catch (error) {
-      console.error('Failed to add wallet info to new chat:', error);
-    } finally {
-      walletUpdateRef.current.isProcessing = false;
-    }
-  }, [initModel, client, getCurrentWalletInfoForPrompt]);
+    //     //  Update ref with current wallet info
+    //     walletUpdateRef.current.previousAddress = walletInfo.address;
+    //     walletUpdateRef.current.previousChainId = walletInfo.chainId;
+    //   }
+    // } catch (error) {
+    //   console.error('Failed to add wallet info to new chat:', error);
+    // } finally {
+    //   walletUpdateRef.current.isProcessing = false;
+    // }
+  }, [initModel, client]);
 
   const onSelectConversation = useCallback(
     async (id: string) => {
@@ -567,39 +345,7 @@ export const useChat = (initValues?: ChatMessage[], initModel?: string) => {
     setActiveChat((prev) => ({ ...prev, model }));
   }, []);
 
-  return useMemo(() => {
-    const walletProviderInfo =
-      walletConnection.isWalletConnected && walletConnection.rdns
-        ? availableProviders.find((p) => p.info.rdns === walletConnection.rdns)
-            ?.info
-        : undefined;
-
-    return {
-      activeChat,
-      conversationHistories,
-      handleNewChat,
-      handleChatCompletion,
-      handleCancel,
-      onSelectConversation,
-      onDeleteConversation,
-      onUpdateConversationTitle,
-      searchableHistory,
-      fetchSearchHistory,
-      toolCallRegistry,
-      walletAddress,
-      handleProviderSelect,
-      availableProviders,
-      walletProviderInfo,
-      isWalletModalOpen,
-      openWalletConnectModal,
-      closeWalletConnectModal,
-      onProviderSelect,
-      changeModel,
-    };
-  }, [
-    walletConnection.isWalletConnected,
-    walletConnection.rdns,
-    availableProviders,
+  return {
     activeChat,
     conversationHistories,
     handleNewChat,
@@ -609,13 +355,8 @@ export const useChat = (initValues?: ChatMessage[], initModel?: string) => {
     onDeleteConversation,
     onUpdateConversationTitle,
     searchableHistory,
+    fetchSearchHistory,
     toolCallRegistry,
-    walletAddress,
-    handleProviderSelect,
-    isWalletModalOpen,
-    openWalletConnectModal,
-    closeWalletConnectModal,
-    onProviderSelect,
     changeModel,
-  ]);
+  };
 };
